[WIP] TST: Clean up testing
There is still some chaos in our test suite, despite recent efforts to
refactor it. This PR tries to improve the situation a bit.

Currently, some tests are expected to fail. This can be due to errors in
this PR but some errors may be genuine bugs. I'll investigate further.

Some of the changes:

- don't add return to pytest.skip, it's not necessary
- avoid self.skipTest, always use pytest.skip
- unused CONFIG_TESTING_KWARGS is removed
- factor out common skipping logic into dedicated functions
- many tests had code like: skip unless it's LoRA or IA³ or ... often,
  new PEFT methdos would actually work but were not added to this list,
  resulting in unnecessary skips