[WIP] TST Enable int4 torchao tests
Trying to train LoRA with int4 torchao used to raise a RuntimeError.
Lately, this error no longer being raised, suggesting that int4 training
is unblocked.

For this to work, we need:

- fbgemm-gpu-genai package from PyTorch (added to Dockerfile)
- transformers has to allow int4 training

The latter is not yet implemented, so this PR is WIP for now.

Moreover, on my machine, I still get an error with int4, namely:

> RuntimeError: cutlass cannot initialize

This could be something specific to my setup and has to be investigated
further.