FIX DeLoRA adapter deletion issue
Currently, adapter deletion raises an error with DeLoRA. The reason is
that the dropout module is called module_dropout, i.e. the prefix
"delora" is not part of the name, which is required for proper working.
This PR renames the attribute to delora_dropout.

The reason why this was not caught is because the corresponding test was
not updated to include DeLoRA. The test was thus also changed.

Note: I came across this issue in #2846 but I wanted to fix it in a
separate PR, as #2846 is probably not going to make it into the next
PEFT release.