FIX: DeLoRA adapter deletion issue (#2853)
Currently, adapter deletion raises an error with DeLoRA. The reason is
that the dropout module is called module_dropout, i.e. the prefix
"delora" is not part of the name, which is required for proper working.
This PR renames the attribute to delora_dropout.

The reason why this was not caught is because the corresponding test was
not updated to include DeLoRA. The test was thus also changed.

Also consolidate merging error message.