Add support for DPO training
In DPO training we have the case that we (potentially) do a forward call first with `disable_adapter`,
then do a forward run with adapters enabled and afterwards do DPO stuff. This is problematic since
we will add gradient checkpointing forward hooks, as long as we detect alora offsets, and the forward
hook code enforces that we only do (forward, backward) calls, (forward, forward) is forbidden since
it would mean we'd register a second pair of hooks for gradient checkpointing.