FIX Correct adapter dtype with bnb weights
Resolves #2889

The reported bug is this: When the base model is quantized with 4bit
bitsandbytes, the adapter weights would be cast to float32, even if
autocast_adapter_dtype=False was passed. This is because the dtype of
the base layer was not correctly determined in that case. This PR now
correctly determines the dtype.

While working on this, I noticed that the peft_model.add_adapter method
was lacking the option to disable autocasting. This was added now and
the tests cover it as well.

TODOs

For LNTuning and OSF, I found that the dtype is not correctly being
applied when it is float16 or bfloat16. As I didn't want to blow up this
PR even more, I decided to skip those methods for now and leave the fix
for those for another time.

Moreover, there is still an issue left with 8bit bnb weights. They don't
have a compute dtype, so at a layer level, it is not possible to
determine what the dtype of the PEFT adapter should be (of course, it
cannot be int8). Therefore, the corresponding tests for 8bit bnb are
x-failing for now. One possible solution could be to pass down the dtype
of the base model (if any) and use that as a fallback. This could be
implemented in a later PR.