FIX: Exploit trust_remote_code in prompt tuning (#2896)
Resolves #2888

This is a test for a hypothetical exploit that would enable
trust_remote_code (and thus RCE) when a user loads a malicious prompt
tuning model. This is because PEFT would just pass the on the
tokenizer_kwargs defined in the prompt tuning config unsanitzed, which
means that if the tokenizer is also malicious, the malicious code would
be executed.

For this exploit to work, a user cannot load a model using
PeftModel.from_pretrained as normal, because the tokenizer is only
loaded in training mode. Although the attacker could set
inference_mode=True in the adapter_config.json, that would still not
work because prompt tuning methods cannot be loaded in inference mode.
Therefore, the only way for the exploit to work would be if the user
manually loads the model.