Add forward compat. for tied_weights_keys dicts
In the future, `_tied_weights_keys` will be mappings (mapping destination as key and source as value).
This will also mean that the semantic of `_tied_weights_keys` will change from "the keys in this list are
tied to the input embedding" to "this mapping defines tying between layers of any type". To this end we'll
limit the scope and provide methods to retrieve input embedding ties.

If the need arises to retrieve more complicated mappings, we can do so at a later point.