TST Add GPU training tests for p/prefix tuning (#2844)
There are already GPU tests for P-tuning and prefix tuning, but none
that involve actual training. This PR adds two training tests (analog to
what we have with LoRA) using transformers Trainer.

This can be important because changes in Trainer, for instance how cache
is handled, can impact the possibility of training with prompt learning
methods.

A recent example for a change is
https://github.com/huggingface/transformers/pull/41585. Although it
doesn't break anything, it shows the potential for breaking in the
Trainer class that could impact these PEFT methods.