TST FIX: Issue with pickle models and caching (#2913)
To avoid rate limits, we use a caching context, which sets offline mode
after the first call, during tests to avoid calling Hugging Face Hub too
frequently. However, we also use a lot of old pickle model
formats (pytorch_model.bin) during testing. These are converted on the
fly on the hub and then loaded using the safetensors format.

This used to work even in offline context but it no longer does (which
is arguably the correct way to handle this). As a consequence, these
models can no longer loaded within the caching context.

This PR switches these models to their safetensors variants, which are
stored in peft-internal-testing. While working on this, we also remove a
few, rarely used model architectures from the decoder model test suite.