FIX Load quantized weights with PEFT mixed model (#2915)
The handling of loading weights with quantized base models by PEFT mixed
model has fallen a bit behind, which resulted in some quantization types
not working. This PR brings it up to date with the normal PEFT model.

As this is a very special case that requires GPU, I think it's overkill
to add a test just for that.