FIX Don't set requires_grad on original module
For modules_to_save, PEFT would enable the gradients on the original
module if the trainable copy was disabled and vice versa. This doesn't
really make sense and was thus removed, even though it is technically a
backwards breaking change.

Also no longer call enable_adapters when _set_adapter is called --
enabling/disabling implies setting the adapter but not vice versa.

Finally, rename _set_adapter_layers to _enable_adapter_layers. This is
just to avoid confusion, as the method deals with enabling, not with
setting.