Fix EETQ GPU Docker image build (#2935)
The basic fix is to install EETQ with `--no-build-isolation` since it imports
torch in setup.py and build isolation seems to shield the environment from the
setup process so that torch is not found.

I took the liberty to convert the `source activate peft` sections to explicit
`conda run` commands to make the use of the conda environment less speculative.

All changes around that are mostly shuffling stuff around / whitespace noise.
The only functionally relevant change is the build isolation.

Co-authored-by: nemo <git@ningu.net>