Fix bug when no active adapter (but still enabled)
These changes were required because
test_load_multiple_adapters_different_modules_to_save was failing with
the recent changes. This is because set_adapter no longer disables the
non-active adapters for auxiliary modules (remember: active adapters and
enabled adapters are two different concepts). As a consequence, it can
now happen that there is an inactive adapter that is *not* disabled.
When dynamically accessing an attribute on ModulesToSave via
__getattr__, e.g. module.weight, there was a bug for exactly this
situation, which results in an error that the attribute cannot be found.
What should actually happen is that since there is no active adapter,
the attribute should be coming from the original module. This bug is
fixed now. (Note that this bugfix is also partly contained in #2893).