[WIP] TST: Clean up testing (#2846)
There is still some chaos in our test suite, despite recent efforts to
refactor it. This PR tries to improve the situation a bit.

Some of the changes:

- for some PEFT methods, adjust some tolerances where necessary instead
of skipping the whole test
- don't add return to pytest.skip, it's not necessary
- avoid self.skipTest, always use pytest.skip
- unused CONFIG_TESTING_KWARGS is removed
- factor out common skipping logic into dedicated functions
- many tests had code like: skip unless it's LoRA or IA³ or ... often,
new PEFT methdos would actually work but were not added to this list,
resulting in unnecessary skips
- make it so that LN Tuning is allowed to run with 0 active adapters
- fix a bug in BOFT resulting in cayley_batch always returning float32
dtype
- transpose delta_weight for WaveFT merging if base layer is Conv1D
- fix UnboundLocalError in C3A

Errors not fixed in this PR:

Layer indexing test is flaky with VBLoRA and VeRA, even with looser
tolerances. Suspiciously, both methods have in common that they share
weights across layers and layer indexing skips some layers. There is
probably something there, but it's not immediately obvious to me. I
checked that this test would already fail on main if not skipped, so I
think it's fine to keep skipping it for these PEFT methods.