FIX Don't set requires_grad on original module (#2936)
For modules_to_save, PEFT would enable the gradients on the original
module if the trainable copy was disabled and vice versa. This doesn't
really make sense and was thus removed, even though it is technically a
backwards breaking change.

Also no longer call enable_adapters when _set_adapter is called --
enabling/disabling implies setting the adapter but not vice versa.

Next, this PR contains changes to AuxiliaryTrainingWrapper.__getattr__
and ModulesToSaveWrapper._hasattr_wrapped. These changes were required
because set_adapter no longer disables the non-active adapters for
auxiliary modules (remember: active adapters and enabled adapters are
two different concepts). As a consequence, it can now happen that there
is an inactive adapter that is not disabled. When dynamically accessing
an attribute on ModulesToSave via __getattr__, e.g. module.weight, there
was a bug for exactly this situation, which results in an error that the
attribute cannot be found. Previously, as setting no adapter to active
implicitly disabled the adapter, the bug was not revealed. What should
actually happen is that since there is no active adapter, the attribute
should be coming from the original module. This bug is fixed now. (Note
that this bugfix is also partly contained in #2893).

Finally, rename _set_adapter_layers to _enable_adapter_layers. This is
just to avoid confusion, as the method deals with enabling, not with
setting.