Test cleaning pytest caches
Currently we run out of disk space when packing the model cache update on
ubuntu 3.10 runs so in an effort to make space, we'll delete the temp dir
pytest uses after the testing has finished.

Also set the CI flag so that test summaries are not truncated which might
be better for inspection of errors.