[WIP] FEAT Add function to convert to LoRA
This adds the possibility to convert a non-LoRA adapter into a LoRA
adapter. Not all LoRA adapters will support this, but many will.

Conversion is not precise, there will be a loss of performance. The
higher the rank, the lower the loss, but also the less efficient the
adapter. Also, for now, this only supports linear layers. Still, this
has some advantages:

- In PEFT, LoRA supports more features than most other methods, e.g.
mixed adapter batches. Thus the converted adapter can be used with those
features.
- Some downstream packages support LoRA adapters, but not other PEFT
methods, e.g. Diffusers. The conversion allows to use a non-LoRA adapter
with those packages.

TODOs:

- Deal with no PEFT config
- Right now, only LoKr support is there, add more
- Documentation

For the future:

- Convert PEFT methods that are non-trivial, like IA³
- Convert layers other than linear layers