CI For transformers main tests, clear disk space
Follow up to #2938. The transformers main tests run into the same issue
of running out of disk space. Apply the same clean up logic as for the
normal CI.

Note that there seem to be options for GH actions to re-use code between
workflows but that seems like overkill for now, so this PR just copies
the step 1:1.