[WIP] Refactor layer initialization
There is a lot of code duplication in the way that adapter layers are
initialized. This means that if, say, a new LoRA config argument is
added, we need updates in the following places:

1. LoraModel._create_and_replace: kwargs dict
2. LoraModel._create_and_replace: update_layer arguments
3. LoraModel._create_and_replace: update_layer signature
4. Linear.__init__: signature (same for each LoRA layer type)
5. Linear.__init__: update_layer arguments
6. resolve_lora_variant: signature if the argument is tied to a LoRA
   variant

It is far too easy to forget something and it's in fact a common mistake
we see, especially by outside contributors.

This PR aims at removing this duplication as much as possible. The crux
is that almost all of these arguments derive from the LoraConfig.
Therefore, the duplication can be reduced by passing the LoraConfig
directly instead of each of its individual arguments.

As is, this change is only implemented by LoRA, which is by far the most
affected by this duplication. However, an argument could be made that
for consistency, the other PEFT methods should be refactored
accordingly.

Possible counter points:

With this change, the used arguments are not listed explicitly anymore.
But since the LoraConfig is a typed dataclass, this replacement does not
stop us from knowing which arguments are available (unlike, say, using
untyped kwargs). So for editing support and readability, this
replacement should not hurt.

Another possible issue with this proposal is that it could easily break
some functionality, as it can be easy to overlook something. Not
everything is covered by tests (some of the more rarely used quantized
layer types, TP layer), so it's possible to miss this.