TST Remove tests that are completely skipped
We have a few tests that just contain pass or will be pytest.skip-ped
completely. These are now replaced by comments. This doesn't really
change the test coverage or improve runtime much, but reduces noise from
the test logs.

For test_custom_models.py, this reduces the amount of tests from 11989
to 10609.